Architecture Diagram and Cost Estimate added.
One of the uploaded documents is the Web application Architecture diagram which is a blue print of the Student record web application that will be created on the cloud using different AWS services. The other uploaded document is the cost-estimation, this is a detailed summary of the estimated cost for creating the web application. The cost-estimate has been generated by the AWS Pricing Calculator.